[Kernels][GPU] Use narrower type for the mha_mask storage type (#49129)
We do not need a UInt here since the value can only hold 3 different
values. Switch to a UInt8

NN_ORIG_REV_ID: 9e5bcc0b09c847bed6850cc55a10cd57d9518849

+++ mojo/kernels/src/nn/mha_mask.mojo
@@ -15,7 +15,7 @@ from utils.numerics import min_or_neg_inf
 struct TileMaskStatus:
     """A tile's masking status."""
 
-    var status: UInt
+    var status: UInt8
 
     # No element is masked.
     alias NO_MASK = Self(0)
